Fix Unit Test failures due to host's ANA support
We recently added NVMe-oF native multipathing support (Change-Id
I3af33c5e43cfb104e436fb785b08fb28b50a031a) but the unit tests assume
that the host they are run on doesn't support it, so when running them
on a system that does support it then 6 tests fail.

This patch just forces this assumption on the class attribute changing
the value of native_multipath_supported from None to False and thus
disabling the detection mechanism.

It also adds a couple of tests for the _set_native_multipath_supported
method that had no tests.

Change-Id: I7d6fff66c5cc27bbd1f5caa737f01bc7784af710