Make ceph job voting (sometimes)
The os-brick-src-devstack-plugin-ceph job's history indicates
that it is too unstable to vote, but at the same time, having
it as non-voting has burned us in merging rbd-related changes
that could be detected by that job.  At the 2022-07-13 cinder
meeting we agreed to compromise by replacing that job with two
jobs, one that votes and runs only on changes that could affect
rbd, and the other that doesn't vote and runs on all other
appropriate changes.

As was the case with the os-brick-src-devstack-plugin-ceph
job, these run only in the check pipeline.

Change-Id: I0777de370ea53212feaff5a791f9b42652bdc7da