Add flake8-logging-format extension
Automated logging format checking was added to cinder in ussuri
by change I1dedc0b31f78f518c, but it wasn't added to os-brick and
a few infelicities have crept into the code.  Enable the extension
and correct the issues.

To be consistent with cinder, we don't enable G200 ("Logging
statement uses exception in arguments").

Change-Id: I650b92bde5509b654d16301ffe50056e1e5ba579