Allow extend_volume for local devices
The local connector is a simple wrapper around an existing device.
Since the device was already set up by some external process, we
can assume that there is also a way to allow extending the volume.

In the extend_volume method, we just return the size of the device, in
the expectation that it was already resized by some external process.

Change-Id: I7d66e9d0b697cd71a72baef5c267c36da78a857e