Use generic testing template
Change If402f9ae0ca06fec0 replaced cycle-specific testing templates
that had to be changed in each project's zuul config file with a
generic template that only needs to be updated in one place, namely,
in the openstack-zuul-jobs repo.

Apparently os-brick didn't get the memo, so we fix that now.

Change-Id: I8202283d5bd5ecede3414fe3e92e95e743df2f67