NVME-oF: Fix to support new "address" in sysfs
Attaching NVMe-oF no longer works in CentosOS 9 stream using nvme 2.4
and libnvme 1.4.

The reason is that the 'address' file in sysfs now has the 'src_addr'
information.

Before we had:
   traddr=127.0.0.1,trsvcid=4420 After:

Now we have:
   traddr=127.0.0.1,trsvcid=4420,src_addr=127.0.0.1

This patch fixes this issue and future proofs for any additional
information that may be added by parsing the contents and searching for
the parts we care: destination address and port.

Closes-Bug: #2035811
Change-Id: I7a33f38fb1b215dd23e2cff3ffa79025cf19def7