NVMe-oF: Fix generation of hostnqn file
In some old nvme-cli versions the NVMe-oF create_hostnqn method fails.

This happens specifically on versions between not having the
show-hostnqn command and having it always return a value. On those
version the command only returns the value present in the file and never
tries to return an idempotent or random value.

This patch adds for that specific case, which is identified by the
stderr message:

   hostnqn is not available -- use nvme gen-hostnqn

Closes-Bug: #2035606
Change-Id: Ic57d0fd85daf358e2b23326022fc471f034b0a2f
(cherry picked from commit 7da3d82773fe68f7ccb2187ceafed012d145799f)