NVMe-oF: Improve hostnqn creation
This patch improves the creation of the /etc/nvme/hostnqn file by using
the system UUID value we usually already know.

This saves us one or two calls to the nvme-cli command and it also
allows older nvme-cli versions that don't have the `show-hostnqn`
command or have it but can only read from file to generate the same
value every time, which may be useful when running inside a container
under some circumstances.

Change-Id: Ib250d213295695390dbdbb3506cb297a86e95218