Update CI for Dalmatian
Only change is to remove the py38 classifier from setup.cfg;
did not modify the python_requires (which is currently >=3.8)

The brick tox.ini doesn't specify specific python versions, so
no modifications there.  We rely on the openstack-python3-jobs
template to specify what versions of python should run in the
gate.

Change-Id: Id27841ba03b915fe3c1cde86273c76e24c521b88