Fix: UTs in fibre channel multipath tests
The unit tests are failing with the following error:

raise mismatch_error
testtools.matchers._impl.MismatchError: !=:
reference = ['multipath -f /dev/mapper/mpatha',
 'tee -a /sys/block/sdb/device/delete',
 'tee -a /sys/block/sdc/device/delete']
actual    = ['multipath -C dm-0',
 'multipath -f /dev/mapper/mpatha',
 'tee -a /sys/block/sdb/device/delete',
 'tee -a /sys/block/sdc/device/delete']

NOTE: Since this is an environment specific failure, it is not
always visible in the failure.

The issue happens because of 2 reasons:
1. we removed the mock of wait_for_mpath_device
introduced in patch[1]
2. we use /dev/sdb as a device name in unit tests

How did the issue occur?
If there is an environment with multipathing enabled and
there exists a device /dev/sdb which is part of a multipath
device, we will execute the ``multipath -C dm-X`` command
and that breaks the assertions as we assume the multipath
-C command shouldn't be running on actual devices in UTs.
There is unlikely to be seen in most of the envs as we
passed the upstream gate in several branches with the same
UTs.

How to fix it?
Reintroduce the mock for wait_for_mpath_device and
find_sysfs_multipath_dm, also add the validations added by
change[2].

Another improvement is also to use `fake-device` like names
instead of actual devices like `/dev/sdb` but that will be a
much bigger change to do here.

[1] https://review.opendev.org/c/openstack/os-brick/+/920516
[2] https://review.opendev.org/c/openstack/os-brick/+/940748

Change-Id: Ib0913d87dfed1ecd79ec4447687eaa0de9248a5c