Make exception raised by Rbd.fileno() more specific
OSError and IOError are the same type. io.UnsupportedOperation is an
exception type deriving from both OSError and ValueError. The
documentation of io.IOBase does suggest to raise UnsupportedOperation
for some things, but not explicitly for fileno(). For fileno(), the
advice is to raise OSError.

This patch changes the exception type to io.UnsupportedOperation
(which still satisfies the requirements stated in io.IOBase). We do
that, because the `requests` library explicitly catches only
io.UnsupportedOperation when calling fileno() while trying to find the
size of a stream passed as body.

Without this patch, it seems that uploading a volume to an image always
fails with a traceback ending in:

```
  File "/usr/local/lib/python3.8/site-packages/requests/models.py", line 526, in prepare_body
    length = super_len(data)
  File "/usr/local/lib/python3.8/site-packages/requests/utils.py", line 150, in super_len
    fileno = o.fileno()
  File "/usr/local/lib/python3.8/site-packages/os_brick/initiator/linuxrbd.py", line 258, in fileno
    raise IOError(_("fileno() not supported by RBD()"))
OSError: fileno() not supported by RBD()
```

Change-Id: I3eae0d241798a5e8984b579db6cc578146052f0e