[stable-only] Cap setuptools <71.0.0
py38 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: I8a08cc142c8dbcb733390b3ec5380e532aece739
Signed-off-by: Bence Romsics <bence.romsics@gmail.com>