LightOS: update discovery client configuration
The LightOS connector uses hardcoded values for discovery client
connection parameters, preventing customization for different
deployment environments and containers with custom networks.

Example of when this is needed:
We have 1 Cinder instance in a Docker container with network 10.10.10.1,
and a discovery-service located at 10.10.10.2.
In the same cluster, we have a Nova instance in a container with
network 20.20.20.1, and a discovery-service at 20.20.20.2.
I need to be able to set the network address for each instance.

Default values are set for backward compatibility.

Change-Id: I19065ad683b9ff8174814d6e8c500178c885bee4
Signed-off-by: yuval <yuval@lightbitslabs.com>